+++ packages/utilities/psammead-assets/src/svgs/navigationIcons.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styled from 'styled-components';
+import styled from '@emotion/styled';
 
 const svgSize = '2.75rem'; // 44px