+++ packages/utilities/psammead-styles/src/global-styles.test.jsx
@@ -1,28 +1,26 @@
 import React from 'react';
-import { render } from '@testing-library/react/pure';
+import { render } from '@testing-library/react';
 import {
   F_REITH_SANS_REGULAR,
   F_REITH_SERIF_MEDIUM,
 } from '@bbc/psammead-styles/fonts';
 import GlobalStyles from './global-styles';
 
-const getGlobalStyles = () => {
-  const rules = document.styleSheets[0].cssRules;
-  return rules.reduce((cssRules, cssRule) => cssRules + cssRule.cssText, '');
-};
+/**
+ * 15/10/20
+ * Revisit snapshot mechanism following Emotion migration.
+ */
 
 describe('global-styles', () => {
   it('should inject normalize to global styles', () => {
     render(<GlobalStyles />);
-    const styles = getGlobalStyles();
-    expect(styles).toMatchSnapshot();
+    expect(document.head).toMatchSnapshot();
   });
 
   it('should handle global-styles with fonts props', () => {
     render(
       <GlobalStyles fonts={[F_REITH_SANS_REGULAR(), F_REITH_SERIF_MEDIUM()]} />,
     );
-    const styles = getGlobalStyles();
-    expect(styles).toMatchSnapshot();
+    expect(document.head).toMatchSnapshot();
   });
 });