+++ packages/components/psammead-live-label/CHANGELOG.md
@@ -3,7 +3,11 @@
 <!-- prettier-ignore -->
 | Version | Description |
 |---------|-------------|
-| 2.0.0 | [PR#3810](https://github.com/bbc/psammead/pull/3810) Migration to Yarn |
+| 3.0.0 | [PR#3810](https://github.com/bbc/psammead/pull/3810) Migration to Yarn |
+| 2.0.2 | [PR#3925](https://github.com/bbc/psammead/pull/3925) Talos - Bump Dependencies - @bbc/psammead-styles, @bbc/psammead-visually-hidden-text |
+| 2.0.1 | [PR#3924](https://github.com/bbc/psammead/pull/3924) Migrate storybook to Emotion |
+| 2.0.0 | [PR#3879](https://github.com/bbc/psammead/pull/3879) Migrate to Emotion |
+| 1.0.3 | [PR#3883](https://github.com/bbc/psammead/pull/3883) Talos - Bump Dependencies - @bbc/psammead-assets, @bbc/psammead-image, @bbc/psammead-styles, @bbc/psammead-visually-hidden-text |
 | 1.0.2 | [PR#3728](https://github.com/bbc/psammead/pull/3728) Tweaks following Storybook v5.3.19 to v6.0.22 update |
 | 1.0.1 | [PR#3709](https://github.com/bbc/psammead/pull/3709) Bump prettier from 2.0.5 to 2.1.1 |
 | 1.0.0 | [PR#3293](https://github.com/BBC/psammead/pull/3293) Remove alpha tag |