+++ packages/components/psammead-media-indicator/CHANGELOG.md
@@ -3,7 +3,13 @@
 <!-- prettier-ignore -->
 | Version | Description |
 | ------- | ----------- |
-| 6.0.0  | [PR#3810](https://github.com/bbc/psammead/pull/3810) Migration to Yarn |
+| 7.0.0  | [PR#3810](https://github.com/bbc/psammead/pull/3810) Migration to Yarn |
+| 6.0.1 | [PR#3929](https://github.com/bbc/psammead/pull/3929) Resolve versioning mismatch |
+| 6.0.0 | [PR#3898](https://github.com/bbc/psammead/pull/3898) Migrate to Emotion |
+| 5.0.4 | [PR#3925](https://github.com/bbc/psammead/pull/3925) Talos - Bump Dependencies - @bbc/psammead-styles, @bbc/psammead-assets |
+| 5.0.3 | [PR#3883](https://github.com/bbc/psammead/pull/3883) Talos - Bump Dependencies - @bbc/psammead-styles, @bbc/psammead-assets |
+| 5.0.2 | [PR#3855](https://github.com/bbc/psammead/pull/3855) Talos - Bump Dependencies - @bbc/gel-foundations |
+| 5.0.1 | [PR#3854](https://github.com/bbc/psammead/pull/3854) Talos - Bump Dependencies - @bbc/psammead-styles |
 | 5.0.0-alpha.0 | [PR#3843](https://github.com/bbc/psammead/pull/3843) Bump Dependencies - @bbc/psammead-styles alpha version |
 | 4.0.17 | [PR#3836](https://github.com/bbc/psammead/pull/3836) Talos - Bump Dependencies - @bbc/psammead-styles |
 | 4.0.16 | [PR#3816](https://github.com/bbc/psammead/pull/3816) Talos - Bump Dependencies |