+++ packages/components/psammead-section-label/CHANGELOG.md
@@ -3,7 +3,13 @@
 <!-- prettier-ignore -->
 | Version | Description |
 |---------|-------------|
-| 7.0.0  | [PR#3810](https://github.com/bbc/psammead/pull/3810) Migration to Yarn |
+| 8.0.0  | [PR#3810](https://github.com/bbc/psammead/pull/3810) Migration to Yarn |
+| 7.0.2 | [PR#3925](https://github.com/bbc/psammead/pull/3925) Talos - Bump Dependencies - @bbc/psammead-styles |
+| 7.0.1 | [PR#3904](https://github.com/bbc/psammead/pull/3904) Emotion migration of titles.jsx |
+| 7.0.0 | [PR#3886](https://github.com/bbc/psammead/pull/3886) Emotion migration |
+| 6.0.3 | [PR#3883](https://github.com/bbc/psammead/pull/3883) Talos - Bump Dependencies - @bbc/psammead-styles |
+| 6.0.2 | [PR#3855](https://github.com/bbc/psammead/pull/3855) Talos - Bump Dependencies - @bbc/gel-foundations |
+| 6.0.1 | [PR#3854](https://github.com/bbc/psammead/pull/3854) Talos - Bump Dependencies - @bbc/psammead-styles |
 | 6.0.0-alpha.0 | [PR#3843](https://github.com/bbc/psammead/pull/3843) Bump Dependencies - @bbc/psammead-styles alpha version |
 | 5.0.17 | [PR#3836](https://github.com/bbc/psammead/pull/3836) Talos - Bump Dependencies - @bbc/psammead-styles |
 | 5.0.16 | [PR#3816](https://github.com/bbc/psammead/pull/3816) Talos - Bump Dependencies |