+++ packages/utilities/web-vitals/CHANGELOG.md
@@ -3,4 +3,5 @@
 | Version | Description |
 |---------|-------------|
 | 2.0.0   | [PR#3810](https://github.com/bbc/psammead/pull/3810) Migration to Yarn |
+| 1.0.1   | [PR#3866](https://github.com/bbc/psammead/pull/3866) Rewrite async function to prevent regeneratorRuntime errors |
 | 1.0.0   | [PR#3784](https://github.com/bbc/psammead/pull/3784) Initial creation of package. |